+++ src/main/java/efim/Item.java
@@ -7,15 +7,15 @@ import java.io.Serializable;
  */
 public class Item implements Serializable {
     public int item;
-    public int utility = 0;
+    public long utility = 0;
 
-    public Item(int i, int u){
+    public Item(int i, long u){
         this.item = i;
         this.utility = u;
     }
 
     public int getItem(){ return this.item; }
-    public int getUtility(){ return this.utility; }
+    public long getUtility(){ return this.utility; }
 
     public void setItem(int item){ this.item = item; }
     public void setUtility(int utility){ this.utility = utility; }
